Move config handler in a separate file
